STCOR-842 refactor away from deprecated getChildContext
`getChildContext` will be deprecated in React v19. `AddContext` can be
eliminated completely; it duplicates the functionality offered by
`StripesContext`.

Refs STCOR-842
