+++ src/AppRoutes.js
@@ -5,7 +5,6 @@ import PropTypes from 'prop-types';
 import { connectFor } from '@folio/stripes-connect';
 
 import { StripesContext } from './StripesContext';
-import AddContext from './AddContext';
 import TitleManager from './components/TitleManager';
 import RouteErrorBoundary from './components/RouteErrorBoundary';
 import { getEventHandlers } from './handlerService';
@@ -62,21 +61,19 @@ const AppRoutes = ({ modules, stripes }) => {
 
         return (
           <StripesContext.Provider value={moduleStripes}>
-            <AddContext context={{ stripes: moduleStripes }}>
-              <ModuleHierarchyProvider module={module.module}>
-                <div id={`${name}-module-display`} data-module={module.module} data-version={module.version}>
-                  <RouteErrorBoundary
-                    escapeRoute={module.home ?? module.route}
-                    moduleName={displayName}
-                    stripes={moduleStripes}
-                  >
-                    <TitleManager page={displayName}>
-                      <ModuleComponent {...props} connect={connect} stripes={moduleStripes} actAs="app" />
-                    </TitleManager>
-                  </RouteErrorBoundary>
-                </div>
-              </ModuleHierarchyProvider>
-            </AddContext>
+            <ModuleHierarchyProvider module={module.module}>
+              <div id={`${name}-module-display`} data-module={module.module} data-version={module.version}>
+                <RouteErrorBoundary
+                  escapeRoute={module.home ?? module.route}
+                  moduleName={displayName}
+                  stripes={moduleStripes}
+                >
+                  <TitleManager page={displayName}>
+                    <ModuleComponent {...props} connect={connect} stripes={moduleStripes} actAs="app" />
+                  </TitleManager>
+                </RouteErrorBoundary>
+              </div>
+            </ModuleHierarchyProvider>
           </StripesContext.Provider>
         );
       }}
